requirements.txt
sentence-transformers
faiss-cpu
requests
beautifulsoup4
python-dotenv
pyyaml
PyMuPDF
prompt_toolkit